Added PW tests - Search - Filters - Logic
